Poisson Surface Reconstruction (Version 3)
EXECUTABLE ARGUMENTS
--in <input points>
This string is the name of the file from which the point set will be
read.
If the --binary flag is not enabled, the file should be an ascii file
with groups of 6, white space delimited, numbers (x-, y-, and
z-coordinates of the point's position, followed by the x-, y-, and
z-coordinates of the point's normal).
If the --binary flag is enabled, the file should be a binary file,
consisting of blocks of 6 32-bit floats (x-, y-, and z-coordinates of
the point's position, followed by the x-, y-, and z-coordinates of the
point's normal).
No information about the number of oriented point samples should be
specified.
--out <output triangle mesh>
This string is the name of the file to which the triangle mesh will be
written. The file is written in PLY format.
[--binary]
If this argument is specified the input samples will be read in as a
binary file.
[--depth <reconstruction depth>]
This integer is the maximum depth of the tree that will be used for
surface reconstruction. Running at depth d corresponds to solving on a
voxel grid whose resolution is no larger than 2^d x 2^d x 2^d. Note that
since the reconstructor adapts the octree to the sampling density, the
specified reconstruction depth is only an upper bound.
The default value for this parameter is 8.
[--scale <scale factor>]
This floating point value specifies the ratio between the diameter of
the cube used for reconstruction and the diameter of the samples'
bounding cube.
The default value is 1.1.
[--solverDivide <solver subdivision depth>]
This integer argument specifies the depth at which a block Gauss-Seidel
solver is used to solve the Laplacian equation. Using this parameter
helps reduce the memory overhead at the cost of a small increase in
reconstruction time. (In practice, we have found that for
reconstructions of depth 9 or higher a subdivide depth of 7 or 8 can
greatly reduce the memory usage.)
The default value is 8.
[--isoDivide <iso-surface extraction subdivision depth>]
This integer argument specifies the depth at which a block iso-surface
extractor should be used to extract the iso-surface. Using this
parameter helps reduce the memory overhead at the cost of a small
increase in extraction time. (In practice, we have found that for
reconstructions of depth 9 or higher a subdivide depth of 7 or 8 can
greatly reduce the memory usage.)
The default value is 8.
[--samplesPerNode <minimum number of samples>]
This floating point value specifies the minimum number of sample points
that should fall within an octree node as the octree construction is
adapted to sampling density. For noise-free samples, small values in the
range [1.0 - 5.0] can be used. For more noisy samples, larger values in
the range [15.0 - 20.0] may be needed to provide a smoother,
noise-reduced, reconstruction.
The default value is 1.0.
[--confidence]
Enabling this flag tells the reconstructor to use the size of the
normals as confidence information. When the flag is not enabled, all
normals are normalized to have unit-length prior to reconstruction.
[--polygonMesh]
Enabling this flag tells the reconstructor to output a polygon mesh
(rather than triangulating the results of Marching Cubes). This option
over-rides the --manifold flag.
[--manifold]
Enabling this flag tells the reconstructor to add the polygon barycenter
when triangulating polygons with more than three vertices.
[--verbose]
Enabling this flag provides a more verbose description of the running
times and memory usages of individual components of the surface
reconstructor.