Screened Poisson Surface Reconstruction (Version 4)
EXECUTABLE ARGUMENTS
--in <input points>
This string is the name of the file from which the point set will be
read.
If the file extension is .ply, the file should be in PLY format, giving
the list of oriented vertices with the x-, y-, and z-coordinates of the
positions encoded by the properties x, y, and z and the x-, y-, and
z-coordinates of the normals encoded by the properties nx, ny, and nz .
If the file extension is .bnpts, the file should be a binary file,
consisting of blocks of 6 32-bit floats: x-, y-, and z-coordinates of
the point's position, followed by the x-, y-, and z-coordinates of the
point's normal. (No information about the number of oriented point
samples should be specified.)
Otherwise, the file should be an ascii file with groups of 6, white
space delimited, numbers: x-, y-, and z-coordinates of the point's
position, followed by the x-, y- and z-coordinates of the point's
normal. (No information about the number of oriented point samples
should be specified.)
[--out <output triangle mesh>]
This string is the name of the file to which the triangle mesh will be
written. The file is written in PLY format.
[--depth <reconstruction depth>]
This integer is the maximum depth of the tree that will be used for
surface reconstruction. Running at depth d corresponds to solving on a
voxel grid whose resolution is no larger than 2^d x 2^d x 2^d. Note that
since the reconstructor adapts the octree to the sampling density, the
specified reconstruction depth is only an upper bound.
The default value for this parameter is 8.
[--minDepth <adaptive octree depth>]
This integer specifies the depth beyond depth the octree will be
adapted. At coarser depths, the octree will be complete, containing all
2^d x 2^d x 2^d nodes.
The default value for this parameter is 5.
[--pointWeight <interpolation weight>]
This floating point value specifies the importants that interpolation of
the point samples is given in the formulation of the screened Poisson
equation.
The results of the original (unscreened) Poisson Reconstruction can be
obtained by setting this value to 0.
The default value for this parameter is 4.
[--threads <number of processing threads>]
This integer specifies the number of threads across which the
reconstruction algorithm should be parallelized.
The default value for this parameter is equal to the numer of (virtual)
processors on the executing machine.
[--scale <scale factor>]
This floating point value specifies the ratio between the diameter of
the cube used for reconstruction and the diameter of the samples'
bounding cube.
The default value is 1.1.
[--solverDivide <solver subdivision depth>]
This integer argument specifies the depth at which a block Gauss-Seidel
solver is used to solve the Laplacian equation. Using this parameter
helps reduce the memory overhead at the cost of a small increase in
reconstruction time. (In practice, we have found that for
reconstructions of depth 9 or higher a subdivide depth of 7 or 8 can
greatly reduce the memory usage.)
The default value is 8.
[--isoDivide <iso-surface extraction subdivision depth>]
This integer argument specifies the depth at which a block iso-surface
extractor should be used to extract the iso-surface. Using this
parameter helps reduce the memory overhead at the cost of a small
increase in extraction time. (In practice, we have found that for
reconstructions of depth 9 or higher a subdivide depth of 7 or 8 can
greatly reduce the memory usage.)
The default value is 8.
[--samplesPerNode <minimum number of samples>]
This floating point value specifies the minimum number of sample points
that should fall within an octree node as the octree construction is
adapted to sampling density. For noise-free samples, small values in the
range [1.0 - 5.0] can be used. For more noisy samples, larger values in
the range [15.0 - 20.0] may be needed to provide a smoother,
noise-reduced, reconstruction.
The default value is 1.0.
[--confidence]
Enabling this flag tells the reconstructor to use the size of the
normals as confidence information. When the flag is not enabled, all
normals are normalized to have unit-length prior to reconstruction.
[--polygonMesh]
Enabling this flag tells the reconstructor to output a polygon mesh
(rather than triangulating the results of Marching Cubes).
[--verbose]
Enabling this flag provides a more verbose description of the running
times and memory usages of individual components of the surface
reconstructor.

--------------------------------------------------------------------------------
CHANGES
Version 3:
The implementation of the --samplesPerNode parameter has been modified
so that a value of "1" more closely corresponds to a distribution with
one sample per leaf node.
The code has been modified to support compilation under MSVC 2010 and
the associated solution and project files are now provided. (Due to a
bug in the Visual Studios compiler, this required modifying the
implementation of some of the bit-shifting operators.)
Version 4:
The code supports screened reconstruction, with interpolation weight
specified through the --pointWeight parameter.
The code has been implemented to support parallel processing, with the
number of threads used for parallelization specified by the --threads
parameter.
The input point set can now also be in PLY format, and the file-type is
determined by the extension, so that the --binary flag is now obsolete.
At depths coarser than the one specified by the value --minDepth the
octree is no longer adaptive but rather complete, simplifying the
prolongation operator.