Screened Poisson Surface Reconstruction (Version 6)
Modified the solver to use Gauss-Seidel relaxation instead of
conjugate-gradients at finer resolution.
Re-ordered the implementation of the solver so that only a windowed
subset of the matrix is in memory at any time, thereby reducing the
memory usage during the solver phase.
Separated the storage of the data associated with the octree nodes from
the topology.